made swiper component reusable
